social_music_api/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pydantic==1.10.15
pymongo==4.8.0
python-decouple==3.8
python-jose==3.3.0
firebase-admin==6.5.0
requests==2.32.3
